Move util/burnham directly into burnham DAG
This is a workaround to deal with skew between the libraries that the Airflow
UI node sees and what's on master.
