+++ dags/burnham.py
@@ -9,7 +9,10 @@ import uuid
 
 from airflow import models
 from airflow.operators import PythonOperator
-from utils.burnham import burnham_bigquery_run, burnham_run, burnham_sensor
+
+from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
+from operators.bq_sensor import BigQuerySQLSensorOperator
+from operators.gcp_container_operator import GKEPodOperator
 
 DAG_OWNER = "rpierzina@mozilla.com"
 DAG_EMAIL = ["telemetry-alerts@mozilla.com", "rpierzina@mozilla.com"]
@@ -45,6 +48,13 @@ WHERE
   AND metrics.string.test_name = "{test_name}"
 """
 
+DEFAULT_GCP_CONN_ID = "google_cloud_derived_datasets"
+DEFAULT_GKE_LOCATION = "us-central1-a"
+DEFAULT_GKE_CLUSTER_NAME = "bq-load-gke-1"
+DEFAULT_GKE_NAMESPACE = "default"
+
+BURNHAM_PLATFORM_URL = "https://incoming.telemetry.mozilla.org"
+
 default_args = {
     "owner": DAG_OWNER,
     "email": DAG_EMAIL,
@@ -54,6 +64,128 @@ default_args = {
     "depends_on_past": False,
     "retries": 0,
 }
+
+
+def burnham_run(
+    task_id,
+    burnham_test_run,
+    burnham_test_name,
+    burnham_missions,
+    burnham_spore_drive=None,
+    gcp_conn_id=DEFAULT_GCP_CONN_ID,
+    gke_location=DEFAULT_GKE_LOCATION,
+    gke_cluster_name=DEFAULT_GKE_CLUSTER_NAME,
+    gke_namespace=DEFAULT_GKE_NAMESPACE,
+    **kwargs
+):
+    """Create a new GKEPodOperator that runs the burnham Docker image.
+
+    :param str task_id:                         [Required] ID for the task
+    :param str burnham_test_run:                [Required] UUID for the test run
+    :param str burnham_test_name:               [Required] Name for the test item
+    :param List[str] burnham_missions:          [Required] List of mission identifiers
+
+    :param Optional[str] burnham_spore_drive:   Interface for the spore-drive technology
+    :param str gcp_conn_id:                     Airflow connection id for GCP access
+    :param str gke_location:                    GKE cluster location
+    :param str gke_cluster_name:                GKE cluster name
+    :param str gke_namespace:                   GKE cluster namespace
+    :param Dict[str, Any] kwargs:               Additional kwargs for GKEPodOperator
+
+    :return: GKEPodOperator
+    """
+    kwargs["name"] = kwargs.get("name", task_id.replace("_", "-"))
+
+    env_vars = {
+        "BURNHAM_PLATFORM_URL": BURNHAM_PLATFORM_URL,
+        "BURNHAM_TEST_RUN": burnham_test_run,
+        "BURNHAM_TEST_NAME": burnham_test_name,
+    }
+
+    if burnham_spore_drive is not None:
+        env_vars["BURNHAM_SPORE_DRIVE"] = burnham_spore_drive
+
+    return GKEPodOperator(
+        task_id=task_id,
+        gcp_conn_id=gcp_conn_id,
+        project_id=GoogleCloudBaseHook(gcp_conn_id=gcp_conn_id).project_id,
+        location=gke_location,
+        cluster_name=gke_cluster_name,
+        namespace=gke_namespace,
+        image="gcr.io/moz-fx-data-airflow-prod-88e0/burnham:latest",
+        image_pull_policy="Always",
+        env_vars=env_vars,
+        arguments=burnham_missions,
+        **kwargs
+    )
+
+
+def burnham_sensor(task_id, sql, gcp_conn_id=DEFAULT_GCP_CONN_ID, **kwargs):
+    """Create a new BigQuerySQLSensorOperator that checks for burnham data.
+
+    :param str task_id:                 [Required] ID for the task
+    :param str sql:                     [Required] SQL for the sensor
+
+    :param str gcp_conn_id:             Airflow connection id for GCP access
+    :param Dict[str, Any] kwargs:       Additional kwargs for BigQuerySQLSensorOperator
+
+    :return: BigQuerySQLSensorOperator
+    """
+    return BigQuerySQLSensorOperator(
+        task_id=task_id,
+        sql=sql,
+        bigquery_conn_id=gcp_conn_id,
+        use_legacy_sql=False,
+        **kwargs
+    )
+
+
+def burnham_bigquery_run(
+    task_id,
+    project_id,
+    burnham_test_run,
+    burnham_test_scenarios,
+    gcp_conn_id=DEFAULT_GCP_CONN_ID,
+    gke_location=DEFAULT_GKE_LOCATION,
+    gke_cluster_name=DEFAULT_GKE_CLUSTER_NAME,
+    gke_namespace=DEFAULT_GKE_NAMESPACE,
+    **kwargs
+):
+    """Create a new GKEPodOperator that runs the burnham-bigquery Docker image.
+
+    :param str task_id:                 [Required] ID for the task
+    :param str project_id:              [Required] Project ID where target table lives
+    :param str burnham_test_run:        [Required] UUID for the test run
+    :param str burnham_test_scenarios:  [Required] Encoded burnham test scenarios
+
+    :param str gcp_conn_id:             Airflow connection id for GCP access
+    :param str gke_location:            GKE cluster location
+    :param str gke_cluster_name:        GKE cluster name
+    :param str gke_namespace:           GKE cluster namespace
+    :param Dict[str, Any] kwargs:       Additional kwargs for GKEPodOperator
+
+    :return: GKEPodOperator
+    """
+    kwargs["name"] = kwargs.get("name", task_id.replace("_", "-"))
+
+    return GKEPodOperator(
+        task_id=task_id,
+        gcp_conn_id=gcp_conn_id,
+        project_id=GoogleCloudBaseHook(gcp_conn_id=gcp_conn_id).project_id,
+        location=gke_location,
+        cluster_name=gke_cluster_name,
+        namespace=gke_namespace,
+        image="gcr.io/moz-fx-data-airflow-prod-88e0/burnham-bigquery:latest",
+        image_pull_policy="Always",
+        arguments=[
+            "--verbose",
+            "--project-id='{}'".format(project_id),
+            "--run-id='{}'".format(burnham_test_run),
+            "--scenarios='{}'".format(burnham_test_scenarios),
+        ],
+        **kwargs
+    )
+
 
 with models.DAG(
     "burnham", schedule_interval="@daily", default_args=default_args,
